models/open_model_5/requirements.txt
numerapi===2.15.1
python-dotenv===1.0.0
pandas===2.1.0
xgboost===2.0.3
pyarrow===13.0.0
psutil===5.9.5